# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: CodeAtlas API
  version: 0.1.0
  description: |
    Current backend API surface for the CodeAtlas V1 local development setup.

    Primary onboarding flow:
    1. Sign in with GitHub through the auth service.
    2. Install the CodeAtlas GitHub App on a personal account or organization.
    3. Claim the returned GitHub App installation for the signed-in CodeAtlas user.
    4. List repositories visible through that installation.
    5. Connect one repository into CodeAtlas for later sync and analytics.
servers:
  - url: http://localhost:8061
    description: Auth service
  - url: http://localhost:8062
    description: Repository service
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Installation:
      type: object
      required: [id, installation_id, status, created_at, updated_at]
      properties:
        id:
          type: integer
          format: int64
        installation_id:
          type: integer
          format: int64
        installed_by_user_id:
          type: integer
          format: int64
          nullable: true
        account_login:
          type: string
          nullable: true
        account_type:
          type: string
          nullable: true
        setup_action:
          type: string
          nullable: true
        status:
          type: string
          example: pending_verification
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    InstallationListResponse:
      type: object
      required: [installations]
      properties:
        installations:
          type: array
          items:
            $ref: '#/components/schemas/Installation'
    ClaimInstallationRequest:
      type: object
      required: [installation_id]
      properties:
        installation_id:
          type: integer
          format: int64
    ClaimInstallationResponse:
      type: object
      required: [installation]
      properties:
        installation:
          $ref: '#/components/schemas/Installation'
...
    Repository:
      type: object
      required: [id, github_repo_id, owner, name, full_name, default_branch, is_private, sync_status, created_at, updated_at]
      properties:
        id:
          type: integer
          format: int64
        github_repo_id:
          type: integer
          format: int64
        owner:
          type: string
        name:
          type: string
        full_name:
          type: string
        default_branch:
          type: string
        is_private:
          type: boolean
        installation_id:
          type: integer
          format: int64
          nullable: true
        webhook_id:
          type: integer
          format: int64
          nullable: true
        sync_status:
          type: string
        last_synced_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    RepositoryListResponse:
      type: object
      required: [repositories]
      properties:
        repositories:
          type: array
          items:
            $ref: '#/components/schemas/Repository'
    RepositoryResponse:
      type: object
      required: [repository, connection_status]
      properties:
        repository:
          $ref: '#/components/schemas/Repository'
        connection_status:
          type: string
          enum: [created, already_connected, updated]
...
    ConnectRepositoryRequest:
      type: object
      required: [github_repo_id, owner, name, default_branch]
      properties:
        github_repo_id:
          type: integer
          format: int64
        owner:
          type: string
        name:
          type: string
        default_branch:
          type: string
        is_private:
          type: boolean
        installation_id:
          type: integer
          format: int64
          nullable: true
...
    InstallURLResponse:
      type: object
      required: [install_url]
      properties:
        install_url:
          type: string
    SetupCallbackResponse:
      type: object
      required: [installation]
      properties:
        installation:
          $ref: '#/components/schemas/Installation'
paths:
...
  /integrations/github/install:
    get:
      summary: Get GitHub App install URL
      security:
        - bearerAuth: []
      responses:
        '200':
          description: GitHub App install link
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstallURLResponse'
        '401':
          description: Missing or invalid JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /integrations/github/setup:
    get:
      summary: Handle GitHub App setup callback
      description: Records the returned installation id from GitHub. May redirect to the frontend with installation details.
      parameters:
        - in: query
          name: installation_id
          required: true
          schema:
            type: integer
            format: int64
        - in: query
          name: setup_action
          schema:
            type: string
      responses:
        '200':
          description: Installation callback recorded directly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetupCallbackResponse'
        '307':
          description: Redirect back to frontend with installation details
        '400':
          description: Missing or invalid installation_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /integrations/github/installations:
    get:
      summary: List GitHub App access linked to the current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Installations for current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstallationListResponse'
        '401':
          description: Missing or invalid JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /integrations/github/installations/claim:
    post:
      summary: Claim a returned GitHub App installation for the current user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimInstallationRequest'
      responses:
        '200':
          description: Installation claimed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimInstallationResponse'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /repos:
    get:
      summary: List repositories already connected for the current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Connected repositories
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepositoryListResponse'
    post:
      summary: Connect a repository directly with provided GitHub metadata
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectRepositoryRequest'
      responses:
        '201':
          description: Repository connected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepositoryResponse'
        '400':
          description: Missing required repository fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...